fix state interfering with shared classic/beta stuff
