+++ web/src/beta/features/Editor/tabs/map/BottomPanel/LayerStyleCard.tsx
@@ -1,4 +1,4 @@
-import { ReactNode, useCallback, useState } from "react";
+import { ReactNode, useCallback, useRef, useState } from "react";
 
 import TextInput from "@reearth/beta/components/fields/common/TextInput";
 import Icon from "@reearth/beta/components/Icon";
@@ -32,26 +32,31 @@ const LayerStyleCard: React.FC<Props> = ({
   const [isHovered, setIsHovered] = useState(false);
   const [isEditing, setIsEditing] = useState(false);
   const [newName, setNewName] = useState(name);
-  const [clickTimeoutId, setClickTimeoutId] = useState<any>(null);
+  const clickTimeoutRef = useRef<NodeJS.Timeout>();
+  const clickedNameCount = useRef(0);
 
   const handleClick = useCallback(() => {
-    if (clickTimeoutId !== null) {
-      clearTimeout(clickTimeoutId);
-      setClickTimeoutId(null);
+    console.log(clickedNameCount.current);
+    if (clickTimeoutRef.current) {
+      clearTimeout(clickTimeoutRef.current);
     }
-    const timeoutId = window.setTimeout(() => {
+    clickTimeoutRef.current = setTimeout(() => {
+      clickedNameCount.current = 0;
       onSelect?.(!selected);
-    }, 200);
-    setClickTimeoutId(timeoutId);
-  }, [clickTimeoutId, onSelect, selected]);
+    }, 100);
+  }, [onSelect, selected]);
 
   const handleNameClick = useCallback(() => {
-    if (clickTimeoutId !== null) {
-      clearTimeout(clickTimeoutId);
-      setClickTimeoutId(null);
+    const newClickCount = clickedNameCount.current + 1;
+    console.log(newClickCount);
+    clickedNameCount.current = newClickCount;
+    // if (clickTimeoutRef.current) {
+    //   clearTimeout(clickTimeoutRef.current);
+    // }
+    if (clickedNameCount.current === 2) {
+      setIsEditing(true);
     }
-    setIsEditing(true);
-  }, [clickTimeoutId]);
+  }, []);
 
   const handleMouseEnter = useCallback(() => {
     setIsHovered(true);
