+++ web/src/classic/components/organisms/EarthEditor/CanvasArea/hooks.ts
@@ -21,6 +21,9 @@ import {
   WidgetAreaAlign,
   ValueType,
 } from "@reearth/classic/gql";
+import { valueTypeToGQL, ValueTypes, valueToGQL, LatLng } from "@reearth/classic/util/value";
+import { config } from "@reearth/services/config";
+import { useLang } from "@reearth/services/i18n";
 import {
   useSceneId,
   useSceneMode,
@@ -32,10 +35,7 @@ import {
   useZoomedLayerId,
   useClock,
   useSelectedWidgetArea,
-} from "@reearth/classic/state";
-import { valueTypeToGQL, ValueTypes, valueToGQL, LatLng } from "@reearth/classic/util/value";
-import { config } from "@reearth/services/config";
-import { useLang } from "@reearth/services/i18n";
+} from "@reearth/services/state";
 
 import {
   convertWidgets,
