+++ web/src/classic/components/organisms/EarthEditor/DataSourcePane/hooks.ts
@@ -12,6 +12,7 @@ import {
   useRemoveDatasetMutation,
   useGetDatasetSchemasWithCountQuery,
 } from "@reearth/classic/gql";
+import { useT, useLang } from "@reearth/services/i18n";
 import {
   useSceneId,
   useNotification,
@@ -19,8 +20,7 @@ import {
   useProject,
   NotificationType,
   useCurrentTheme,
-} from "@reearth/classic/state";
-import { useT, useLang } from "@reearth/services/i18n";
+} from "@reearth/services/state";
 
 export default () => {
   const t = useT();
