+++ web/src/classic/components/organisms/EarthEditor/PropertyPane/hooks.ts
@@ -23,6 +23,8 @@ import {
   WidgetSectionType,
   WidgetZoneType,
 } from "@reearth/classic/gql";
+import { valueTypeToGQL, Camera, toGQLSimpleValue, valueToGQL } from "@reearth/classic/util/value";
+import { useLang } from "@reearth/services/i18n";
 import {
   useSelected,
   useRootLayerId,
@@ -34,9 +36,7 @@ import {
   useSelectedBlock,
   useWidgetAlignEditorActivated,
   useSelectedWidgetArea,
-} from "@reearth/classic/state";
-import { valueTypeToGQL, Camera, toGQLSimpleValue, valueToGQL } from "@reearth/classic/util/value";
-import { useLang } from "@reearth/services/i18n";
+} from "@reearth/services/state";
 
 import useQueries, { Mode as RawMode } from "./hooks-queries";
