+++ web/src/classic/components/organisms/EarthEditor/core/CanvasArea/hooks.ts
@@ -25,6 +25,14 @@ import {
   ValueType,
 } from "@reearth/classic/gql";
 import {
+  valueTypeToGQL,
+  type ValueTypes,
+  valueToGQL,
+  type LatLng,
+} from "@reearth/classic/util/value";
+import { config } from "@reearth/services/config";
+import { useLang } from "@reearth/services/i18n";
+import {
   useSceneId,
   useSceneMode,
   useIsCapturing,
@@ -34,15 +42,7 @@ import {
   useWidgetAlignEditorActivated,
   useZoomedLayerId,
   useSelectedWidgetArea,
-} from "@reearth/classic/state";
-import {
-  valueTypeToGQL,
-  type ValueTypes,
-  valueToGQL,
-  type LatLng,
-} from "@reearth/classic/util/value";
-import { config } from "@reearth/services/config";
-import { useLang } from "@reearth/services/i18n";
+} from "@reearth/services/state";
 
 import {
   convertWidgets,
