--- web/src/classic/components/organisms/Notification/index.tsx
@@ -1,20 +1,0 @@
-import React from "react";
-
-import MoleculeNotificationBanner from "@reearth/classic/components/atoms/Notification";
-
-import useHooks from "./hooks";
-
-const NotificationBanner: React.FC = () => {
-  const { visible, notification, setModal, resetNotification } = useHooks();
-
-  return (
-    <MoleculeNotificationBanner
-      visible={visible}
-      notification={notification}
-      setModal={setModal}
-      resetNotification={resetNotification}
-    />
-  );
-};
-
-export default NotificationBanner;
