+++ web/src/classic/components/organisms/Settings/Workspace/Asset/hooks.ts
@@ -2,7 +2,7 @@ import { useEffect } from "react";
 import { useNavigate } from "react-router-dom";
 
 import assetHooks from "@reearth/classic/components/organisms/Common/AssetContainer/hooks";
-import { useWorkspace, useProject } from "@reearth/classic/state";
+import { useWorkspace, useProject } from "@reearth/services/state";
 
 export type Params = {
   workspaceId: string;
