+++ web/src/classic/components/pages/Authentication/hooks.ts
@@ -3,9 +3,9 @@ import { useCallback, useEffect, useMemo } from "react";
 import { useLocation, useNavigate } from "react-router-dom";
 
 import { useGetTeamsQuery } from "@reearth/classic/gql";
-import { useWorkspace, useNotification, useUserId } from "@reearth/classic/state";
 import { useAuth, useCleanUrl } from "@reearth/services/auth";
 import { useT } from "@reearth/services/i18n";
+import { useWorkspace, useNotification, useUserId } from "@reearth/services/state";
 
 // TODO: move hooks to molecules (page components should be thin)
 export default () => {
