+++ web/src/services/state/index.ts
@@ -1,4 +1,5 @@
 import { atom, useAtom } from "jotai";
+import { atomWithStorage } from "jotai/utils";
 
 import type {
   ComputedFeature,
@@ -6,6 +7,8 @@ import type {
   LayerSelectionReason,
 } from "@reearth/beta/lib/core/Map";
 import type { Camera } from "@reearth/beta/utils/value";
+import { Clock } from "@reearth/classic/components/molecules/Visualizer/Plugin/types";
+import { ProjectType } from "@reearth/types";
 
 export { default as useSetError, useError } from "./gqlErrorHandling";
 
@@ -103,8 +106,13 @@ export const useSelectedWidgetArea = () => useAtom(selectedWidgetArea);
 const notification = atom<Notification | undefined>(undefined);
 export const useNotification = () => useAtom(notification);
 
-// Not sure we need belowNot sure we need belowNot sure we need belowNot sure we need below
-// Not sure we need belowNot sure we need belowNot sure we need belowNot sure we need below
+// --------------------------------------------------------------------------------------------------------------------------------
+// --------------------------------------------------------------------------------------------------------------------------------
+// Not sure we need below in Beta. Currently being used in classic.Not sure we need below in Beta. Currently being used in classic.
+// Not sure we need below in Beta. Currently being used in classic.Not sure we need below in Beta. Currently being used in classic.
+// --------------------------------------------------------------------------------------------------------------------------------
+// --------------------------------------------------------------------------------------------------------------------------------
+
 const isCapturing = atom<boolean>(false);
 export const useIsCapturing = () => useAtom(isCapturing);
 
@@ -123,3 +131,57 @@ export const useCurrentTheme = () => useAtom(currentTheme);
 
 const workspace = atom<Workspace | undefined>(undefined);
 export const useWorkspace = () => useAtom(workspace);
+
+const sceneId = atom<string | undefined>(undefined);
+export const useSceneId = () => useAtom(sceneId);
+
+const rootLayerId = atom<string | undefined>(undefined);
+export const useRootLayerId = () => useAtom(rootLayerId);
+
+export type Selected =
+  | { type: "scene" }
+  | {
+      type: "layer";
+      layerId: string;
+      featureId?: string;
+      layerSelectionReason?: LayerSelectionReason;
+    }
+  | { type: "widgets" }
+  | { type: "cluster"; clusterId: string }
+  | { type: "widget"; widgetId?: string; pluginId: string; extensionId: string }
+  | { type: "dataset"; datasetSchemaId: string };
+
+const selected = atom<Selected | undefined>(undefined);
+export const useSelected = () => useAtom(selected);
+
+const camera = atom<Camera | undefined>(undefined);
+export const useCamera = () => useAtom(camera);
+
+const clock = atom<Clock | undefined>(undefined);
+export const useClock = () => useAtom(clock);
+
+const userId = atomWithStorage<string | undefined>("userId", undefined);
+export const useUserId = () => useAtom(userId);
+
+export type Project = {
+  id: string;
+  name: string;
+  sceneId?: string;
+  isArchived?: boolean;
+  projectType?: ProjectType;
+};
+
+const project = atom<Project | undefined>(undefined);
+export const useProject = () => useAtom(project);
+
+const unselectProject = atom(null, (_get, set) => {
+  set(project, undefined);
+  set(workspace, undefined);
+  set(sceneId, undefined);
+  set(selected, undefined);
+  set(selectedBlock, undefined);
+  set(camera, undefined);
+  set(isCapturing, false);
+  set(sceneMode, "3d");
+});
+export const useUnselectProject = () => useAtom(unselectProject)[1];
